feat: catalog item details panel — tabs, pricing, API, tools, visual polish
