--- libs/catalog/src/components/CatalogItemDetails/CatalogOverview.module.scss
@@ -1,34 +1,0 @@
-// All colors via CSS custom properties so consumers can theme freely.
-
-.sectionHeading {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.specLabel {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.specValueText {
-  color: var(--text-primary, #eef1f7);
-}
-
-.specValueTrue {
-  color: var(--text-primary, #eef1f7);
-}
-
-.specValueFalse {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.specCheckIcon {
-  flex-shrink: 0;
-  color: var(--text-accent-secondary, #37babc);
-}
-
-.specRow {
-  border-radius: 4px;
-}
-
-.specRowAlt {
-  background: var(--bg-layer-1, #0d1425);
-}
