--- libs/catalog/src/components/CatalogItemDetails/CatalogPricing.module.scss
@@ -1,19 +1,0 @@
-.sectionHeading {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.row {
-  border-radius: 4px;
-}
-
-.rowAlt {
-  background: var(--bg-layer-1, #0d1425);
-}
-
-.label {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.value {
-  text-align: start;
-}
