--- libs/catalog/src/components/CatalogItemDetails/CatalogTools.module.scss
@@ -1,42 +1,0 @@
-.divider {
-  border-top: 1px solid var(--stroke-tertiary, #1a2035);
-}
-
-.description {
-  color: var(--text-secondary, #9fa6bd);
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  border: 1px solid var(--stroke-secondary, #242c42);
-  border-radius: 6px;
-  overflow: hidden;
-}
-
-.th {
-  color: var(--text-secondary, #9fa6bd);
-  background: var(--bg-layer-1, #0d1425);
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-  font-weight: 500;
-}
-
-.tr {
-  &:nth-child(even) {
-    background: var(--bg-layer-1, #0d1425);
-  }
-}
-
-.td {
-  color: var(--text-primary, #e4e6f0);
-  border-top: 1px solid var(--stroke-secondary, #242c42);
-}
-
-.code {
-  font-family: 'Fira Code', 'Cascadia Code', 'Courier New', monospace;
-  background: none;
-  border: none;
-  padding: 0;
-  color: inherit;
-}
