+++ libs/catalog/src/components/Details/CatalogApiDetails.module.scss
