+++ libs/catalog/src/components/Details/CatalogApiDetails.tsx
