+++ libs/catalog/src/components/Details/DetailsPanel.module.scss
@@ -7,6 +7,11 @@
 
 .panel {
   background: var(--bg-layer-0, #000000);
+  border-inline-start-color: var(--stroke-secondary, #242c42);
+}
+
+.divider {
+  border-block-end: 1px solid var(--stroke-secondary, #242c42);
 }
 
 .version {
@@ -24,16 +29,18 @@
   scrollbar-color: var(--stroke-secondary, #242c42) transparent;
 }
 
-.skeletonLine {
-  background: var(--stroke-secondary, #242c42);
-}
-
 .introCaption {
   color: var(--text-secondary, #9fa6bd);
 }
 
 .featuredTag {
-  background: var(--cat-featured-tag-bg, var(--controls-bg-accent-secondary-alpha-active, #37babc5c));
-  border-color: var(--cat-featured-tag-border, var(--stroke-accent-secondary, #37babc));
+  background: var(
+    --cat-featured-tag-bg,
+    var(--controls-bg-accent-secondary-alpha-active, #37babc5c)
+  );
+  border-color: var(
+    --cat-featured-tag-border,
+    var(--stroke-accent-secondary, #37babc)
+  );
   color: var(--cat-featured-tag-text, var(--text-accent-secondary, #37babc));
 }
