+++ libs/catalog/src/components/Details/Summary/Limits.module.scss
@@ -1,10 +1,4 @@
 .tag {
-  display: inline-flex;
-  align-items: center;
-  padding: 2px 8px;
-  border-radius: 4px;
-  font-weight: 500;
-
   &--FREE { background: var(--bg-accent-secondary, rgba(83, 194, 91, 0.12)); color: var(--text-accent-secondary, #53c25b); }
   &--FEATURED { background: var(--bg-warning-secondary, rgba(238, 200, 64, 0.12)); color: var(--text-warning, #eec840); }
   &--BY_REQUEST { background: var(--bg-layer-2, #1a2035); color: var(--text-secondary, #9fa6bd); }
@@ -12,24 +6,12 @@
   &--DEPRECATED { background: var(--bg-layer-2, #1a2035); color: var(--text-error, #f2655a); }
 }
 
-.badgeImage {
-  height: 20px;
-  width: auto;
-  object-fit: contain;
-}
-
 .progressTrack {
-  height: 6px;
-  border-radius: 3px;
   background: var(--bg-layer-2, #1a2035);
-  overflow: hidden;
 }
 
 .progressFill {
-  height: 100%;
-  border-radius: 3px;
   background: var(--text-accent-primary, #6374ff);
-  transition: width 0.3s ease;
 }
 
 .resetLabel {
