+++ libs/catalog/src/components/Details/Summary/Limits.tsx
@@ -1,8 +1,9 @@
 import { mergeClasses } from '@epam/ai-dial-chat-shared';
+import { DialTag } from '@epam/ai-dial-ui-kit';
 import { FC } from 'react';
-import type { CatalogItemSummary } from '../../models/entity-summary';
-import { EntityTag } from '../../types/entity-tag';
-import styles from './CatalogSummary.module.scss';
+import type { CatalogItemSummary } from '../../../models/entity-summary';
+import { EntityTag } from '../../../types/entity-tag';
+import styles from './Limits.module.scss';
 
 const TAG_LABELS: Record<EntityTag, string> = {
   [EntityTag.Free]: 'Free',
@@ -12,8 +13,8 @@ const TAG_LABELS: Record<EntityTag, string> = {
   [EntityTag.Deprecated]: 'Deprecated',
 };
 
-/** Props for `CatalogSummary`. */
-export interface CatalogSummaryProps {
+/** Props for `Limits`. */
+export interface LimitsProps {
   /** Summary data to render. */
   summary: CatalogItemSummary;
   /** Label shown above the daily-limit progress bar. Default: `'Daily limit'`. */
@@ -27,7 +28,7 @@ export interface CatalogSummaryProps {
 }
 
 /** Renders the header summary block: entity tag, badge image, and daily-limit progress bar. */
-export const CatalogSummary: FC<CatalogSummaryProps> = ({
+export const Limits: FC<LimitsProps> = ({
   summary,
   dailyLimitLabel = 'Daily limit',
   tagClassName = 'dial-tiny-text',
@@ -36,28 +37,29 @@ export const CatalogSummary: FC<CatalogSummaryProps> = ({
 }) => {
   const { tag, badgeImageUrl, dailyLimit } = summary;
 
+  console.log('Summary:', summary);
+  console.log('Tag:', tag);
   return (
     <div className="flex flex-col gap-3">
       {((tag != null && tag !== EntityTag.Featured) ||
         badgeImageUrl != null) && (
         <div className="flex flex-wrap items-center gap-2">
           {tag != null && tag !== EntityTag.Featured && (
-            <span
+            <DialTag
+              label={TAG_LABELS[tag]}
               className={mergeClasses(
                 styles.tag,
-                styles[`tag--${tag.toLowerCase()}`],
+                styles[`tag--${tag}`],
                 tagClassName,
               )}
-            >
-              {TAG_LABELS[tag]}
-            </span>
+            />
           )}
           {badgeImageUrl != null && (
             <img
               src={badgeImageUrl}
               alt=""
               aria-hidden="true"
-              className={styles.badgeImage}
+              className="h-5 w-auto object-contain"
             />
           )}
         </div>
@@ -76,14 +78,20 @@ export const CatalogSummary: FC<CatalogSummaryProps> = ({
             )}
           </div>
           <div
-            className={styles.progressTrack}
+            className={mergeClasses(
+              'h-1.5 overflow-hidden rounded-[3px]',
+              styles.progressTrack,
+            )}
             role="progressbar"
             aria-valuenow={dailyLimit.used}
             aria-valuemin={0}
             aria-valuemax={dailyLimit.total}
           >
             <div
-              className={styles.progressFill}
+              className={mergeClasses(
+                'h-full rounded-[3px] [transition:width_0.3s_ease]',
+                styles.progressFill,
+              )}
               style={{
                 width: `${Math.min(100, (dailyLimit.used / dailyLimit.total) * 100)}%`,
               }}
